iwater/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
flask-smorest==0.42.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-jwt-extended==4.6.0
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
waitress==2.1.2
